fix: #963 config not work
